+++ samples/poke-sample/src/main/java/sample/basic/Break.java
@@ -14,16 +14,10 @@
  * limitations under the License.
  */
 
-include ':library:android-mvc'
-include ':library:android-mvc-test'
-include ':library:android-mvc-core'
-include ':library:poke'
+package sample.basic;
 
-include ':extension:service-core'
-include ':extension:service-mediastore'
-
-include 'samples:benchmark'
-
-include 'samples:simple-mvp:app'
-include 'samples:simple-mvp:core'
-include 'samples:poke-sample'
+public class Break {
+    public void slow(Car car) {
+        car.setSpeed(car.getSpeed() - 1);
+    }
+}