+++ samples/poke-sample/src/main/java/sample/scope/Tourist.java
@@ -14,16 +14,13 @@
  * limitations under the License.
  */
 
-include ':library:android-mvc'
-include ':library:android-mvc-test'
-include ':library:android-mvc-core'
-include ':library:poke'
+package sample.scope;
 
-include ':extension:service-core'
-include ':extension:service-mediastore'
+import javax.inject.Inject;
 
-include 'samples:benchmark'
+import sample.basic.Car;
 
-include 'samples:simple-mvp:app'
-include 'samples:simple-mvp:core'
-include 'samples:poke-sample'
+public class Tourist {
+    @Inject
+    Car car;
+}